add email confirmation for social account

+++ database/migrations/2023_03_22_001038_create_products_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->decimal('price', 10, 2);
+            $table->double('price', 10, 2);
             $table->integer('stock');
             $table->string('image')->nullable();
             $table->timestamps();
